Move generic classes to slang-plugin and plugin classes to new plugin package
