+++ slang-plugin/src/main/java/org/sonarsource/slang/plugin/ParseException.java
@@ -17,27 +17,29 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-package org.sonarsource.slang.kotlin;
+package org.sonarsource.slang.plugin;
 
-import org.sonar.api.config.Configuration;
-import org.sonar.api.resources.AbstractLanguage;
+import org.sonarsource.slang.api.TextPointer;
+import javax.annotation.CheckForNull;
+import javax.annotation.Nullable;
 
-public class KotlinLanguage extends AbstractLanguage {
+public class ParseException extends RuntimeException {
 
-  private Configuration configuration;
+  private final TextPointer position;
 
-  public KotlinLanguage(Configuration configuration) {
-    super(KotlinPlugin.KOTLIN_LANGUAGE_KEY, KotlinPlugin.KOTLIN_LANGUAGE_NAME);
-    this.configuration = configuration;
+  public ParseException(String message, @Nullable TextPointer position) {
+    super(message);
+    this.position = position;
   }
 
-  @Override
-  public String[] getFileSuffixes() {
-    String[] suffixes = configuration.getStringArray(KotlinPlugin.KOTLIN_FILE_SUFFIXES_KEY);
-    if (suffixes == null || suffixes.length == 0) {
-      suffixes = KotlinPlugin.KOTLIN_FILE_SUFFIXES_DEFAULT_VALUE.split(",");
-    }
-    return suffixes;
+  public ParseException(String message) {
+    super(message);
+    this.position = null;
+  }
+
+  @CheckForNull
+  public TextPointer getPosition() {
+    return position;
   }
 
 }
