+++ sonar-kotlin-plugin/pom.xml
@@ -83,7 +83,7 @@
         <configuration>
           <!-- This value shows up in the SonarQube GUI -->
           <pluginName>SonarKotlin</pluginName>
-          <pluginClass>org.sonarsource.slang.kotlin.KotlinPlugin</pluginClass>
+          <pluginClass>org.sonarsource.slang.kotlin.plugin.KotlinPlugin</pluginClass>
           <skipDependenciesPackaging>true</skipDependenciesPackaging>
           <sonarLintSupported>true</sonarLintSupported>
           <addMavenDescriptor>false</addMavenDescriptor>
