+++ sonar-kotlin-plugin/src/main/java/org/sonarsource/slang/kotlin/KotlinCodeVerifier.java
@@ -38,6 +38,7 @@ import org.jetbrains.kotlin.psi.KtPostfixExpression;
 import org.jetbrains.kotlin.psi.KtPrefixExpression;
 import org.jetbrains.kotlin.psi.KtStringTemplateExpression;
 import org.jetbrains.kotlin.psi.KtThisExpression;
+import org.sonarsource.slang.plugin.ParseException;
 
 public class KotlinCodeVerifier implements CodeVerifier {
   private static final List<String> KDOC_TAGS = Arrays.asList(
