+++ sonar-kotlin-plugin/src/main/java/org/sonarsource/slang/kotlin/KotlinTreeVisitor.java
@@ -117,6 +117,7 @@ import org.jetbrains.kotlin.psi.KtWhenCondition;
 import org.jetbrains.kotlin.psi.KtWhenEntry;
 import org.jetbrains.kotlin.psi.KtWhenExpression;
 import org.jetbrains.kotlin.psi.KtWhileExpression;
+import org.sonarsource.slang.plugin.ParseException;
 
 import static org.sonarsource.slang.api.LoopTree.LoopKind.DOWHILE;
 import static org.sonarsource.slang.api.LoopTree.LoopKind.FOR;
