+++ sonar-kotlin-plugin/src/main/java/org/sonarsource/slang/kotlin/plugin/KotlinLanguage.java
@@ -17,29 +17,27 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-package org.sonarsource.slang.kotlin;
+package org.sonarsource.slang.kotlin.plugin;
 
-import org.sonarsource.slang.api.TextPointer;
-import javax.annotation.CheckForNull;
-import javax.annotation.Nullable;
+import org.sonar.api.config.Configuration;
+import org.sonar.api.resources.AbstractLanguage;
 
-public class ParseException extends RuntimeException {
+public class KotlinLanguage extends AbstractLanguage {
 
-  private final TextPointer position;
+  private Configuration configuration;
 
-  public ParseException(String message, @Nullable TextPointer position) {
-    super(message);
-    this.position = position;
+  public KotlinLanguage(Configuration configuration) {
+    super(KotlinPlugin.KOTLIN_LANGUAGE_KEY, KotlinPlugin.KOTLIN_LANGUAGE_NAME);
+    this.configuration = configuration;
   }
 
-  public ParseException(String message) {
-    super(message);
-    this.position = null;
-  }
-
-  @CheckForNull
-  public TextPointer getPosition() {
-    return position;
+  @Override
+  public String[] getFileSuffixes() {
+    String[] suffixes = configuration.getStringArray(KotlinPlugin.KOTLIN_FILE_SUFFIXES_KEY);
+    if (suffixes == null || suffixes.length == 0) {
+      suffixes = KotlinPlugin.KOTLIN_FILE_SUFFIXES_DEFAULT_VALUE.split(",");
+    }
+    return suffixes;
   }
 
 }
