+++ sonar-kotlin-plugin/src/main/java/org/sonarsource/slang/kotlin/plugin/KotlinPlugin.java
@@ -17,7 +17,7 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-package org.sonarsource.slang.kotlin;
+package org.sonarsource.slang.kotlin.plugin;
 
 import org.sonarsource.slang.externalreport.androidlint.AndroidLintRulesDefinition;
 import org.sonarsource.slang.externalreport.androidlint.AndroidLintSensor;
