+++ sonar-kotlin-plugin/src/main/java/org/sonarsource/slang/kotlin/plugin/KotlinProfileDefinition.java
@@ -17,14 +17,14 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-package org.sonarsource.slang.kotlin;
+package org.sonarsource.slang.kotlin.plugin;
 
 import org.sonar.api.server.profile.BuiltInQualityProfilesDefinition;
 import org.sonarsource.analyzer.commons.BuiltInQualityProfileJsonLoader;
 
 public class KotlinProfileDefinition implements BuiltInQualityProfilesDefinition {
 
-  public static final String PATH_TO_JSON = "org/sonar/l10n/kotlin/rules/kotlin/Sonar_way_profile.json";
+  private static final String PATH_TO_JSON = "org/sonar/l10n/kotlin/rules/kotlin/Sonar_way_profile.json";
 
   @Override
   public void define(Context context) {
